Address David's review comments
