+++ keras_cv/models/segmentation/__init__.py
@@ -14,6 +14,6 @@
 
 from keras_cv.models.segmentation.deeplab_v3_plus import DeepLabV3Plus
 from keras_cv.models.segmentation.segformer import SegFormer
-from keras_cv.models.segmentation.segment_anything import MaskDecoder
-from keras_cv.models.segmentation.segment_anything import PromptEncoder
+from keras_cv.models.segmentation.segment_anything import SAMMaskDecoder
+from keras_cv.models.segmentation.segment_anything import SAMPromptEncoder
 from keras_cv.models.segmentation.segment_anything import TwoWayTransformer
